CLEAN AND TRANSFORM 2016 ELECTORAL DATA
The raw electoral data from 2016 is organized a bit differently than the 2006 and 2011 files, respectively. This script performs cleans the data (similar to the cleaning scripts for 2006 and 2011) with the relevant tweaks to make it work for the 2016 data. Final output is "data_2016_clean_ward.RDS", and RDS file that has electoral returns cleaned and aggregated to the 2016 ward boundaries.